refactor(lowlevel/envelope): move time source trait and default impl out
