+++ internal/lowlevel/envelope/time_source.go
@@ -1,0 +1,17 @@
+package envelope
+
+import (
+	"time"
+)
+
+type TimeSource interface {
+	GetCurrentTimestamp() time.Time
+}
+
+type DefaultTimeSource struct{}
+
+var _ TimeSource = DefaultTimeSource{}
+
+func (DefaultTimeSource) GetCurrentTimestamp() time.Time {
+	return time.Now()
+}
